Cleaned up listpattern validator
